% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devPTIpack-package.R
\docType{package}
\name{devPTIpack-package}
\alias{devPTIpack}
\alias{devPTIpack-package}
\title{devPTIpack: Priority Targeting Index analysis toolkit}
\description{
Computes, visualises, and explores Priority Targeting Indices
(PTI) over administrative geographies. The package ships a Shiny
application launched via [launch_pti()] together with the
Shiny-free pipeline orchestrator [run_pti_pipeline()].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://worldbank.github.io/devPTIpack/}
  \item \url{https://stunning-engine-gzjwl6l.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Koichi Ito \email{koichi19960406@gmail.com}

Authors:
\itemize{
  \item Eduard Bukin \email{ebukin@gmail.com}
}

Other contributors:
\itemize{
  \item Takaaki Masaki \email{tmasaki@worldbank.org} [contributor]
  \item Arden Finn \email{afinn1@worldbank.org} [reviewer]
}

}
\keyword{internal}
